Mandatory_Assignment_2/requirements.txt
absl-py==2.1.0
astunparse==1.6.3
box2d-py==2.3.8
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
cloudpickle==3.0.0
contourpy==1.3.0
cycler==0.12.1
flatbuffers==24.3.25
fonttools==4.54.1
gast==0.6.0
google-auth==2.35.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.66.2
gym==0.26.2
gym-notices==0.0.8
h5py==3.12.1
idna==3.10
keras==2.14.0
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.9.2
ml-dtypes==0.2.0
numpy==1.23.5
oauthlib==3.2.2
opt_einsum==3.4.0
packaging==24.1
pillow==10.4.0
protobuf==4.25.5
pyasn1==0.6.1
pyasn1_modules==0.4.1
pygame==2.6.1
pyparsing==3.1.4
PyQt5==5.15.11
PyQt5-Qt5==5.15.15
PyQt5_sip==12.15.0
python-dateutil==2.9.0.post0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow-cpu==2.14.0
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.4.0
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.4
wrapt==1.14.1
